+++ Exceptions/IServiceExecutor.cs
@@ -1,16 +1,15 @@
-﻿namespace SPOrchestratorAPI.Exceptions
+﻿namespace SPOrchestratorAPI.Exceptions;
+
+/// <summary>
+/// Interfaz que define un servicio para ejecutar acciones de manera reactiva y manejar errores.
+/// </summary>
+public interface IServiceExecutor
 {
     /// <summary>
-    /// Interfaz que define un servicio para ejecutar acciones de manera reactiva y manejar errores.
+    /// Ejecuta una acción de manera reactiva y maneja errores, retornando un observable de tipo <typeparamref name="T"/>.
     /// </summary>
-    public interface IServiceExecutor
-    {
-        /// <summary>
-        /// Ejecuta una acción de manera reactiva y maneja errores, retornando un observable de tipo <typeparamref name="T"/>.
-        /// </summary>
-        /// <typeparam name="T">Tipo de dato que emitirá el observable.</typeparam>
-        /// <param name="action">Func que retorna un IObservable de tipo <typeparamref name="T"/>.</param>
-        /// <returns>Un observable que emite objetos de tipo <typeparamref name="T"/> o lanza excepciones.</returns>
-        IObservable<T> ExecuteAsync<T>(Func<IObservable<T>> action);
-    }
+    /// <typeparam name="T">Tipo de dato que emitirá el observable.</typeparam>
+    /// <param name="action">Func que retorna un IObservable de tipo <typeparamref name="T"/>.</param>
+    /// <returns>Un observable que emite objetos de tipo <typeparamref name="T"/> o lanza excepciones.</returns>
+    IObservable<T> ExecuteAsync<T>(Func<IObservable<T>> action);
 }
