+++ Exceptions/ResourceNotFoundException.cs
@@ -1,7 +1,6 @@
-﻿namespace SPOrchestratorAPI.Exceptions
-{
-    /// <summary>
-    /// Excepción personalizada para indicar que un recurso no fue encontrado en la base de datos.
-    /// </summary>
-    public class ResourceNotFoundException(string message) : Exception(message);
-}
+﻿namespace SPOrchestratorAPI.Exceptions;
+
+/// <summary>
+/// Excepción personalizada para indicar que un recurso no fue encontrado en la base de datos.
+/// </summary>
+public class ResourceNotFoundException(string message) : Exception(message);
