+++ SPOrchestratorAPI.csproj
@@ -18,4 +18,8 @@
       <PackageReference Include="System.Reactive" Version="6.0.1" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Controllers\ServicioConfiguracionControllers\" />
+    </ItemGroup>
+
 </Project>
